Self-help changes

+++ src/components/App.tsx
@@ -24,6 +24,8 @@ import ImportActivity from "./Researcher/ActivityList/ImportActivity"
 import PreventPage from "./PreventPage"
 import { sensorEventUpdate } from "./BottomMenu"
 import TwoFA from "./TwoFA"
+import demo_db from "../demo_db.json"
+import self_help_db from "../self_help_db.json"
 
 function ErrorFallback({ error }) {
   const [trace, setTrace] = useState([])
@@ -129,6 +131,11 @@ function AppRouter({ ...props }) {
   const serverAddressFro2FA = ["api-staging.lamp.digital", "api.lamp.digital"]
 
   useEffect(() => {
+    if (localStorage.getItem("demo_mode") === "try_it") {
+      LAMP.initializeDemoDB(demo_db)
+    } else if (localStorage.getItem("demo_mode") === "self_help") {
+      LAMP.initializeDemoDB(self_help_db)
+    }
     let query = window.location.hash.split("?")
     if (!!query && query.length > 1) {
       let src = Object.fromEntries(new URLSearchParams(query[1]))["src"]
@@ -276,7 +283,9 @@ function AppRouter({ ...props }) {
   }, [state])
 
   let reset = async (identity?: any) => {
-    Service.deleteUserDB()
+    if (identity?.id != "selfHelp@demo.lamp.digital") {
+      Service.deleteUserDB()
+    }
     Service.deleteDB()
     if (typeof identity === "undefined" && LAMP.Auth._type === "participant") {
       await sensorEventUpdate(null, (state.identity as any)?.id ?? null, null)
